Check for root nginx config

+++ web/build-docker-image.sh
@@ -29,6 +29,12 @@ docker login --username=${NEXUS_LOGIN_USER} --password=${NEXUS_LOGIN_PASSWORD} $
 aws s3 cp s3://mti-ci-artifacts/${PROJECT}/${FULL_VERSION}/${PACKAGE_NAME}.tar.gz ${PACKAGE_NAME}.tar.gz
 tar -zxf ./${PACKAGE_NAME}.tar.gz --directory ./pkg/app
 cp .ci/web/Dockerfile ./pkg/Dockerfile
-cp .ci/web/nginx.conf ./pkg/nginx.conf
+
+if test -f "nginx.conf"; then
+        cp nginx.conf ./pkg/nginx.conf
+    else
+        cp .ci/web/nginx.conf ./pkg/nginx.conf
+fi
+
 docker build --tag ${IMAGE_TAG} ./pkg
 docker push ${IMAGE_TAG}
